backend/requirements.txt
Django>=5.2.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
Pillow>=10.4.0


